+++ chronos_npm_package/server/server.js
@@ -20,11 +20,13 @@ app.get('/random', (req, res) => {
 }
 );
 
-app.get('api/updateDashboard', async (req, res) => {
-    const { graphType } = req.body;
+app.post('/api/updateDashboard', async (req, res) => {
+    const { graphType, metric, token } = req.body;
+    console.log(graphType, metric, token);
     console.log('updateDashboard endpoint hit');
-    const datasource = await utilities.getGrafanaDatasource();
-    await utilities.updateGrafanaDashboard(graphType, token, metric, datasource);
+    const datasource = await utilities.updateGrafanaDatasource();
+    //await utilities.updateGrafanaDashboard(graphType, token, metric, datasource);
+    console.log('Dashboard Updated');
     return res.status(200).send('Dashboard Updated');
 });
